+++ .github/workflows/webflow_percy.yml
@@ -1,0 +1,77 @@
+name: Webflow Percy
+
+on:
+  workflow_dispatch:
+    inputs:
+      environment:
+        type: choice
+        description: Select the environment
+        options:
+        - https://deriv-staged.webflow.io/
+        - https://webflow.deriv.com/ 
+      suite:
+        type: choice
+        description: Run full or rerun
+        options:
+        - full
+        - rerun
+      rerun from :
+        type: string
+        required: false
+        description: In case of rerun select test cases (Optional)
+        options:
+        - webflow/homePage
+jobs:
+  webflow-percy-run:
+    runs-on: ubuntu-latest
+    strategy:
+        fail-fast: false # https://github.com/cypress-io/github-action/issues/48
+
+    steps:
+    - name: Checkout external repository with Cypress tests
+      uses: actions/checkout@v4
+      with:
+        repository: deriv-com/e2e-deriv-com # Replace with your repository name
+    - name: Install percy cli
+        run: |
+          npm i @percy/cli@1.27.3
+          echo "Installed Percy CLI version: $(npx percy --version)"
+    - name: Install percy cypress
+        run: |
+          npm install --save-dev @percy/cypress
+          npm install cypress --save-dev
+          echo "Percy Cypress version: $(npx percy --version)"
+          echo "Cypress version: $(npx cypress --version)" 
+    - name: Run Percy  
+      if: ${{ github.event.inputs.suite = 'full' && github.event.inputs.rerun = '' }}
+      # Uses the official Cypress GitHub action https://github.com/cypress-io/github-action
+      uses: cypress-io/github-action@1b70233146622b69e789ccdd4f9452adc638d25a
+      with:
+        spec: cypress/e2e/webflow/*.js
+      env:
+        TEST_SUITE: ${{ github.event.inputs.suite }}
+        # For recording and parallelization to work you must set your CYPRESS_RECORD_KEY
+        # in GitHub repo → Settings → Secrets → Actions
+        # CYPRESS_RECORD_KEY: ${{ secrets.CYPRESS_RECORD_KEY }}
+        # Creating a token https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token
+        # GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        # Set Base Url from client_payload.
+        CYPRESS_BASE_URL: "${{ github.event.inputs.environment }}"
+        # Send PR details to Cypress test run
+        COMMIT_INFO_MESSAGE: Manual Run of "${{ github.event.inputs.suite }} tests" on "${{ github.event.inputs.environment }}"
+    
+    - name: Run Percy rerun 
+      if: ${{ github.event.inputs.suite == 'rerun' && github.event.inputs.rerun != '' }}
+      # Uses the official Cypress GitHub action https://github.com/cypress-io/github-action
+      uses: cypress-io/github-action@1b70233146622b69e789ccdd4f9452adc638d25a
+      with:
+        spec: cypress/e2e/${{ env.TEST_SUITE }}/*.js
+      env:
+        TEST_SUITE: ${{ github.event.inputs.suite }}
+        # Creating a token https://docs.github.com/en/authentication/keeping-your-account-and-data-secure/creating-a-personal-access-token
+        # GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+        # Set Base Url from client_payload.
+        CYPRESS_BASE_URL: "${{ github.event.inputs.environment }}"
+        # Send PR details to Cypress test run
+        COMMIT_INFO_MESSAGE: Manual Run of "${{ github.event.inputs.suite }} tests" on "${{ github.event.inputs.environment }}"
+
