+++ cypress/support/e2e.js
@@ -93,4 +93,14 @@ Cypress.Commands.add('c_checkAllPlatformLinks',(urlDetails)=>{
   })
 })
 
+Cypress.Commands.add('c_captureScreenshot', (pageName, viewport) => {
+  if (viewport == 'mobile') {
+      cy.percySnapshot(pageName, { widths: [400] })
+  } else if (viewport == 'medium') {
+      cy.percySnapshot(pageName, { widths: [800] })
+  } else if (viewport == 'desktop') {
+      cy.percySnapshot(pageName, { widths: [3072] })
+  }
+})
+
